requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[email]
boto3
pandas
pyarrow
python-dotenv
gunicorn

